Movie Recommendation System
The System is designed to enhance the user's movie-watching experience by providing personalized movie suggestions based on their past interactions, preferences, or viewing history. The system analyzes the user's previously watched or selected movies and recommends new films that align with their tastes. The goal is to reduce the time spent searching for movies and increase satisfaction with the recommendations.